Switch to using args rather than stdin.
This not only fixes a lot of problems with Go in general, but it also fixes problems with goimports not being able to find dependencies because they are inside of `go mod`.  With the latest patches to `goimports` there is support for `go mod` now, so it's best to switch to args so that `goimports` has the context it needs.